DEFAULT_LEVELS moved out and world_generator introduced.
