+++ aimmo-game/simulation/map_generator.py
@@ -10,7 +10,7 @@ from simulation.direction import ALL_DIRECTIONS
 from simulation.location import Location
 from simulation.world_map import WorldMap
 from simulation.world_map import world_map_static_spawn_decorator
-from simulation.world_map import DEFAULT_LEVEL_SETTINGS
+from simulation.maps.world_generator import DEFAULT_LEVEL_SETTINGS
 
 from simulation.custom_map import BaseGenerator
 from simulation.custom_map import BaseLevelGenerator
