+++ aimmo-game/simulation/world_map.py
@@ -6,19 +6,9 @@ from simulation.world_state import MapFeature
 from simulation.pickups import ALL_PICKUPS
 from simulation.location import Location
 from simulation.maps.cell import Cell
+from simulation.maps.world_generator import DEFAULT_LEVEL_SETTINGS
 
 LOGGER = getLogger(__name__)
-
-
-DEFAULT_LEVEL_SETTINGS = {
-    'TARGET_NUM_CELLS_PER_AVATAR': 0,
-    'TARGET_NUM_SCORE_LOCATIONS_PER_AVATAR': 0,
-    'SCORE_DESPAWN_CHANCE': 0,
-    'TARGET_NUM_PICKUPS_PER_AVATAR': 0,
-    'PICKUP_SPAWN_CHANCE': 0,
-    'NO_FOG_OF_WAR_DISTANCE': 1000,
-    'PARTIAL_FOG_OF_WAR_DISTANCE': 1000,
-}
 
 
 class WorldMap(object):
